// fringe/types/api/createPerformanceDto.ts
/**
 * Generated by orval v7.7.0 🍺
 * Do not edit manually.
 * Fringe API
 * API for Fringe application
 * OpenAPI spec version: v1
 */
import type { SeatingType } from './seatingType';

export interface CreatePerformanceDto {
  showId: number;
  performanceDate: string;
  startTime: string;
  endTime: string;
  seatingType?: SeatingType;
  soldOut?: boolean;
  cancel?: boolean;
  active?: boolean;
}