// fringe/types/api/updatePerformanceDto.ts
/**
 * Generated by orval v7.7.0 🍺
 * Do not edit manually.
 * Fringe API
 * API for Fringe application
 * OpenAPI spec version: v1
 */

export interface UpdatePerformanceDto {
  performanceDate: string;
  startTime: string;
  endTime: string;
  soldOut?: boolean;
  cancel?: boolean;
  active?: boolean;
}